Generic Content Type plugin
