+++ gems/aws-sdk-core/lib/aws-sdk-core/rest.rb
@@ -6,7 +6,6 @@ require_relative 'rest/request/builder'
 require_relative 'rest/request/endpoint'
 require_relative 'rest/request/headers'
 require_relative 'rest/request/querystring_builder'
-require_relative 'rest/request/content_type'
 require_relative 'rest/response/body'
 require_relative 'rest/response/headers'
 require_relative 'rest/response/parser'
