+++ gems/aws-sdk-s3/lib/aws-sdk-s3/presigner.rb
@@ -199,7 +199,7 @@ module Aws
         req.handlers.remove(Aws::S3::Plugins::S3Signer::LegacyHandler)
         req.handlers.remove(Aws::Plugins::Sign::Handler)
         req.handlers.remove(Seahorse::Client::Plugins::ContentLength::Handler)
-        req.handlers.remove(Aws::Rest::ContentTypeHandler)
+        req.handlers.remove(Seahorse::Client::Plugins::ContentType::Handler)
 
         req.handle(step: :send) do |context|
           # if an endpoint was not provided, force secure or insecure
